osmを新規作成し、classをosmに移動
